+++ src/libraries/Servo.h
@@ -40,7 +40,7 @@ class Servo
 {
 public:
   Servo();
-  uint8_t attach(int16_t pin);           // attach the given pin to the next free channel, sets pinMode, returns channel number or 0 if failure
+  uint8_t attach(int pin);           // attach the given pin to the next free channel, sets pinMode, returns channel number or 0 if failure
   uint8_t attach(int pin, int min, int max); // as above but also sets min and max values for writes. 
   void detach();
   void write(int val);               // if value is < 200 its treated as an angle, otherwise as pulse width in microseconds
